% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiWaveAnalisysParallel.R
\name{availableFeatures}
\alias{availableFeatures}
\title{availableFeatures}
\usage{
availableFeatures()
}
\value{
A \code{data.frame} containing the name of the characteristics and their
abbreviations for use in the code. For example, to use variances and
correlations, the vector c("Var", "Cor") will be used.
}
\description{
Print the available features for the \code{\link{MultiWaveAnalysis}} and
\code{\link{StepDiscrim}}
}
\examples{
availableFeatures()

}
\seealso{
\itemize{
\item \code{\link{MultiWaveAnalysis}}
\item \code{\link{StepDiscrim}}
\item \code{\link{StepDiscrimV}}
}
}
