% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiWaveAnalisysParallel.R
\name{availableFilters}
\alias{availableFilters}
\title{availableFilters}
\usage{
availableFilters()
}
\value{
A \code{data.frame} containing all supported filters
}
\description{
Print the available filters for the wave analysis
}
\examples{
availableFilters()

}
\seealso{
\itemize{
\item \code{\link{MultiWaveAnalysis}}
}
}
