% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFuncts.R
\name{filterParameters}
\alias{filterParameters}
\title{filterParameters}
\usage{
filterParameters(data, accuracy)
}
\arguments{
\item{data}{Collection of parameter tests obtained with the
\code{\link{testFilters}} function.}

\item{accuracy}{Determines from which precision the elements will be
selected. For example, a value of 0.9 will select all
combinations with a precision equal to or greater than 0.9.}
}
\value{
The parameter combinations that have achieved an accuracy greater
than or equal to that indicated in the accuracy parameter.
}
\description{
This function allows to filter the results obtained by the
\code{\link{testFilters}} function according to the precision achieved.
}
\examples{
\donttest{
load(system.file("extdata/ECGExample.rda",package = "TSEAL"))
# The dataset has the first 5 elements of class 1
# and the last 5 of class 2.
labels <- c(rep(1, 8), rep(2, 8))
result <- testFilters(ECGExample, labels, features=c("var","cor"),
          filters= c("haar","d4"), maxvars = 3, trainSize = 1)
filterResult <- filterParameters(result, 0.9)
}

}
\seealso{
\itemize{
\item \code{\link{testFilters}}
}
}
