% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Discriminant.R
\name{KFCV}
\alias{KFCV}
\title{K-Fold Cross Validation (KFCV)}
\usage{
KFCV(data, ...)
}
\arguments{
\item{data}{Starting data to generate the validation. It can be either the
raw data, or a previously generated MultiWaveAnalysis object.}

\item{...}{Additional arguments}
}
\value{
Not return value, used as generic function
}
\description{
This function performs the K-Fold Cross Validation (KFCV) process
with different types of input parameters.
}
\seealso{
\itemize{
\item \code{\link{KFCV.array}}
\item \code{\link{KFCV.MultiWaveAnalysis}}
}
}
