% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Discriminant.R
\name{LOOCV}
\alias{LOOCV}
\title{Leave-One-Out Cross Validation}
\usage{
LOOCV(data, ...)
}
\arguments{
\item{data}{Starting data to generate the validation. It can be either the
raw data, or a previously generated MultiWaveAnalysis object.}

\item{...}{Additional arguments}
}
\value{
Not return value, used as generic function
}
\description{
This function performs the Leave-One-Out Cross Validation (LOOCV) process
with different types of input parameters.
}
\seealso{
\itemize{
\item \code{\link{LOOCV.array}}
\item \code{\link{LOOCV.MultiWaveAnalysis}}
}
}
