% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Discriminant.R
\name{LOOCV.WaveAnalisys}
\alias{LOOCV.WaveAnalisys}
\title{LOOCV}
\usage{
\method{LOOCV}{WaveAnalisys}(MWA, grps, method, returnClassification = FALSE)
}
\arguments{
\item{MWA}{WaveAnalysis object obtained with MultiWaveAnalisys function}

\item{grps}{labeled vector that classify the observations.}

\item{method}{Selected method for discrimination. Valid options "Linear" "Quadratic"}
}
\description{
Performs a leave-one-cross-validation (LOOCV) method
}
