% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiWaveAnalisysParallel.R
\name{MultiVaweAnalisys}
\alias{MultiVaweAnalisys}
\title{Generate a MultiWave analisys}
\usage{
MultiVaweAnalisys(
  XSeries1,
  XSeries2,
  f,
  lev = 0,
  Var = TRUE,
  Cor = TRUE,
  nCores = 0
)
}
\arguments{
\item{XSeries1}{Sample from the population 1 (dim x length x cases)}

\item{XSeries2}{Sample from the population 2 (dim x length x cases)}

\item{f}{Selected wavelet filter for the analisys. To view all the avaivle filters type ...}

\item{lev}{Wavelet descomposition level by default is selcted using the "conservative" strategy. See chooseLevel function.}

\item{Var}{a boolean that determines if the analissy contains the variances.}

\item{Cor}{a boolean that determines if the analissy contains the correlations.}

\item{nCores}{determines the number of processes that will be used in the function, by default it uses all but one of the system cores.}
}
\value{
An object of class WaveAnalisys
}
\description{
Generate a MultiWave analisys
}
