% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Discriminant.R
\name{trainModel}
\alias{trainModel}
\title{Generate a Discriminant Model}
\usage{
trainModel(data, ...)
}
\arguments{
\item{data}{Starting data to generate a discriminator (linear or quadratic).
This starting data can be either the raw data, or a MultiWaveAnalysis
object generated earlier.}

\item{...}{Additional arguments}
}
\value{
A trained discriminant model
}
\description{
This function allows training of a discriminant model using different inputs
}
\seealso{
\itemize{
\item \code{\link{trainModel.array}}
\item \code{\link{trainModel.MultiWaveAnalysis}}
}
}
