% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Discriminant.R
\name{trainModelMWA}
\alias{trainModelMWA}
\title{train Model
Generates a prediction model from an already generated "WaveAnalysis".}
\usage{
trainModelMWA(MWA, groups, method)
}
\arguments{
\item{MWA}{WaveAnalysis object obtained with MultiWaveAnalisys function}

\item{groups}{labeled vector that classify the observations.}

\item{method}{Selected method for discrimination. Valid options "Linear" "Quadratic"}
}
\value{
a prediction model based on selected method.
}
\description{
train Model
Generates a prediction model from an already generated "WaveAnalysis".
}
