% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiWaveAnalisysParallel.R
\name{chooseLevel}
\alias{chooseLevel}
\title{Select the dwt level of decomposition based on wavelet filter, data series length and a user choice}
\usage{
chooseLevel(choice, f, N)
}
\arguments{
\item{choice}{Valid values:
- "Conservative" : L < log2 ( N / (L - 1) + 1)
- "max" : L <= log2(N)
- "superMax" <= log2(1.5 * N)}

\item{f}{Wavelet transform filter name. For avaible filters use ...}

\item{N}{Number of observations.}
}
\value{
Number of level of descomposition based in selection criteria
}
\description{
Select the dwt level of decomposition based on wavelet filter, data series length and a user choice
}
\examples{
 lev <- chooseLevel("conservative","haar",8)
}
\references{
Percival, D. B. and A. T. Walden (2000) Wavelet Methods for
  Time Series Analysis. Cambridge: Cambridge University Press.
}
