% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Discriminant.R
\name{KFCV}
\alias{KFCV}
\title{KFCV}
\usage{
KFCV(MWA, grps, method, k, seed = 10)
}
\arguments{
\item{MWA}{WaveAnalysis object obtained with MultiWaveAnalisys function}

\item{grps}{labeled vector that classify the observations.}

\item{method}{Selected method for discrimination. Valid options "Linear" "Quadratic"}

\item{k}{k parameter for the K-fold method.}

\item{seed}{seed for random selection of k groups}
}
\value{
A vector containing one confusion matrix object for each group
}
\description{
Performs k-fold cross-validation where groups are chosen randomly.
 In case the value k is not divisor of the number of observations the last
 group will have nobs mod k observations.
}
