% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StepDiscrim.R
\name{StepDiscrim}
\alias{StepDiscrim}
\title{Stepwise discriminant}
\usage{
StepDiscrim(
  WVC,
  grps,
  maxvars,
  features = c("Var", "Cor", "IQR", "PE", "DM"),
  nCores = 0,
  pos = FALSE
)
}
\arguments{
\item{WVC}{WaveAnalisys object obtained with MultiVaweAnalisys function}

\item{grps}{labeled vector that classify the observations.}

\item{maxvars}{The number of desired values.}

\item{nCores}{determines the number of processes that will be used in the function, by default it uses all but one of the system cores.}

\item{Var}{Determines if the algorithm take in account the variances.
For use this option, the provided WaveAnalisys has to have variances}

\item{Cor}{Determines if the algorithm take in account the correlations.
For use this option, the provided WaveAnalisys has to have correlations}
}
\value{
A WaveAnalisys object with the maxvars most discriminant variables,
 and a importance vector that determines the importance order of the selected variables.
}
\description{
Stepwise discriminant analysis to determine the best subset of variables.
Introduces variables so as to maximize at each step the Lawley-Hotelling
trace (=Rao's V).  This measure is proportional to the mean Mahalanobis distance.
}
\details{
Based on StepDiscrim of R.E. Strauss
}
