% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFuncts.R
\name{testFilters}
\alias{testFilters}
\title{testFilters}
\usage{
testFilters(
  XSeries1,
  XSeries2,
  var,
  filters = c("haar", "d4", "d6", "d8", "la8"),
  lev = 0
)
}
\arguments{
\item{XSeries1}{Samples from the population 1 (dim x length x cases)}

\item{XSeries2}{Samples from the population 2 (dim x length x cases)}

\item{var}{Maximum number of selectable variables by stepwise discriminant}

\item{filters}{Vector que contiene los filtros a probar. Para ver los filtros disponibles ejecuta avaibleFilters()}

\item{lev}{Wavelet descomposition level, by default is selcted using the "conservative" strategy. See \code{\link{chooseLevel}} function.}
}
\value{
A list that each element contains:
\itemize{
\item CM: confusion matrix with a particular configuration using LOOCV
\item Vars: a boolean that determines if the variances have been taken into account in the variable selection process.
\item Cors: a boolean that determines if the Correlations have been taken into account in the variable selection process.
\item NVars: the total numbers of variables have been taken into account in the clasification process
\item Method: Type of classifier used.
\item Filter: filter used in the MultiWave analisys process
}
}
\description{
This function performs a test with a series of filters defined by the user,
for the maximum number of variables determined. This function can be used to
compare the performance of different filters with a different number of
variables to be considered and the differences between a linear and a quadratic discriminant.
}
\seealso{
\itemize{
\item \code{\link{LOOCV}}
\item \code{\link{MultiVaweAnalisys}}
\item \code{\link{StepDiscrim}}
}
}
