% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Discriminant.R
\name{trainModel.array}
\alias{trainModel.array}
\title{Train model
Generates a prediction model from training data.}
\usage{
\method{trainModel}{array}(
  XSeries1,
  XSeries2,
  f,
  method,
  maxvars = 0,
  Vstep = 0,
  lev = 0,
  features = c("Var", "Cor", "IQR", "PE", "DM"),
  nCores = 0
)
}
\arguments{
\item{XSeries1}{Sample from the population 1 (dim x length x cases)}

\item{XSeries2}{Sample from the population 2 (dim x length x cases)}

\item{f}{Selected filter for the MODWT (to see the available filters use the function avaibleFilters)}

\item{method}{Selected method for the discriminant. Valid values "Linear" "Quadratic"}

\item{maxvars}{maximun number of variables included by the stepDiscrim algorithm (Note that if you defined this, can not define VStep)}

\item{Vstep}{Minimum value of V above which all other variables are considered
irrelevant and therefore will not be included. (Note that if you defined
this, can not defined maxvars) For more information see StepDiscrim documentation}

\item{lev}{Determines the number of decomposition levels for MODWT (by default the optimum is calculated).}

\item{Var}{Determines whether variances are to be taken into account for the classifier.}

\item{Cor}{Determines whether correlations are to be taken into account for the classifier.}
}
\value{
a prediction model object
}
\description{
It generates a prediction model starting from the training data, which must
be provided in 2 groups depending on their classification. The method first
obtains the variances and correlations using MODWT, the f filter is applied
with a number of levels lev. Then a subset of all the generated features
will be obtained by means of a stepwise discriminant, which can be driven
by a maximum number of features or by a minimum metric to be met. Finally,
the selected prediction model is trained with the subset obtained.
}
\seealso{
\itemize{
\item \code{\link{StepDiscrim}}
\item \code{\link{StepDiscrimV}}
}
}
